[MobileLantern] ✨ Enables dynamic app download button
Implements dynamic loading of the app download button based on a JSON configuration file.
This allows enabling or disabling the download button without code deployments.
The download button is now dynamically updated based on the `mobile-lantern-download.json` file, which contains information about the availability of the APK, download URL, and version.

Adds a "Coming Soon" button as a fallback.
The download information is reloaded every 30 seconds.